Graql and Concept fixes (#54)
## What is the goal of this PR?

To fix a number of Graql and Concept issues, namely:

- Cyclic dependencies in a type graph cause stack overflow errors
- `value` can be defined (and is ignored) on a non-attribute type
- `owns @key` can be defined on a type with instances that don't have that key, breaking graph integrity
- abstract relations can not be defined without roles
- an integer can't be assigned to a `double` attribute
- nothing interesting happens when doing `insert $x isa thing`
- `undefine person sub thing` throws an error
- undefining related roles 'relates', playable roles 'plays', and attribute ownerships 'owns' are all allowed, even if they have existing instances
- unsetPlays throws an error if the role wasn't actually played
- attempting to undefine the wrong regex throws an exception
- behaviour of putRule is weird

## What are the changes implemented in this PR?

- Fix cyclic ownership, cyclic type hierarchy and cyclic roleplaying throwing stack overflow errors
- Add proper cyclic type hierarchy checking
- don't allow `value` to be defined on a non-attribute type
- don't allow `owns @key` to be set with no prior `owns` if instances exist
- allow relation types to have no roles if they are abstract
- allow an integer literal to be assigned to a `double` attribute
- throw an exception on running `insert $x isa thing`
- allow things to be undefined using `sub thing`
- undefining related roles 'relates', playable roles 'plays', and attribute ownerships 'owns' are no longer allowed if they have existing instances
- unsetPlays is now a no-op if the role wasn't actually played
- attempting to undefine the wrong regex is now a no-op
- putRule updates a rule if it exists, else creates it.

+++ concept/type/impl/RoleTypeImpl.java
@@ -32,6 +32,7 @@ import java.util.List;
 import java.util.stream.Stream;
 
 import static grakn.core.common.exception.ErrorMessage.TypeRead.TYPE_ROOT_MISMATCH;
+import static grakn.core.common.exception.ErrorMessage.TypeWrite.INVALID_UNDEFINE_RELATES_HAS_INSTANCES;
 import static grakn.core.common.exception.ErrorMessage.TypeWrite.ROOT_TYPE_MUTATION;
 import static grakn.core.graph.util.Encoding.Vertex.Type.Root.ROLE;
 
@@ -71,12 +72,17 @@ public class RoleTypeImpl extends TypeImpl implements RoleType {
     }
 
     void sup(final RoleType superType) {
-        super.superTypeVertex(((RoleTypeImpl) superType).vertex);
+        super.setSuperTypeVertex(((RoleTypeImpl) superType).vertex);
     }
 
     @Override
     public String getScope() {
         return vertex.scope();
+    }
+
+    @Override
+    public String getScopedLabel() {
+        return vertex.scopedLabel();
     }
 
     @Nullable
@@ -112,7 +118,14 @@ public class RoleTypeImpl extends TypeImpl implements RoleType {
 
     @Override
     public void delete() {
+        if (getInstances().findAny().isPresent()) {
+            throw new GraknException(INVALID_UNDEFINE_RELATES_HAS_INSTANCES.message(getScopedLabel()));
+        }
         vertex.delete();
+    }
+
+    private Stream<RoleImpl> getInstances() {
+        return instances(RoleImpl::of);
     }
 
     @Override
